add parthenon_loops
