+++ example/particle_tracers/particle_tracers.cpp
@@ -477,7 +477,7 @@ TaskCollection ParticleDriver::MakeTaskCollection(BlockList_t &blocks, int stage
     auto set = tl.AddTask(recv, parthenon::cell_centered_bvars::SetBounds<any>, mc1);
 
     if (pmesh->multilevel) {
-      tl.AddTask(set, parthenon::cell_centered_refinement::RestrictPhysicalBounds,
+      tl.AddTask(set, parthenon::refinement::RestrictPhysicalBounds,
                  mc1.get());
     }
   }
