+++ src/globals.hpp
@@ -42,13 +42,13 @@ extern SparseConfig sparse_config;
 extern Real receive_boundary_buffer_timeout;
 extern Real current_task_runtime_sec;
 
-namespace cell_centered_refinement {
+namespace refinement {
 // Communication buffers are packed into a `BufferInfo_t` object.
 // if the size of this object is greater than min_num_bufs,
 // hierarchical parallelism is used for prolongation/restriction.
 // otherwise one kernel per buffer is launched.
 extern int min_num_bufs;
-} // namespace cell_centered_refinement
+} // namespace refinement
 
 } // namespace Globals
 } // namespace parthenon
