+++ src/parthenon_manager.cpp
@@ -138,7 +138,7 @@ ParthenonStatus ParthenonManager::ParthenonInitEnv(int argc, char *argv[]) {
       pinput->GetOrAddReal("parthenon/time", "recv_bdry_buf_timeout_sec", -1.0);
 
   // set boundary comms buffer switch trigger
-  Globals::cell_centered_refinement::min_num_bufs =
+  Globals::refinement::min_num_bufs =
       pinput->GetOrAddReal("parthenon/mesh", "refinement_in_one_min_nbufs", 64);
 
   return ParthenonStatus::ok;
